Add ErrorMessageContainer and Icon_Vinyl
Created ErrorMessageContainer and Icon_Vinyl.
ErrorMessageContainer holds error messages for the text inputs. The messages are passed as props.

The Icon_Vinyl shows a 24px x24px version of the big vinyl disc shown in the Hero component. This icon is also animated to spin. It's purpose is to be used with the Toastify

Update SearcBar and TrackList text inputs to show an Indicator with the string "Required"