Finished styling
Hero component: now the title has a radial gradient starting from white at the center and finishing with base-200 on the edges, to make sure the Jammming string always has a white background to accentuate the vinyl font.

Searchbar: now the text matches the other ones, with text-primary instead of text-black

SearchResults: now the texts match the other ones, with text-primary instead of text-black
The add button now has a tooltip to indicate its action

SortRadios: changed the color of the labels to text-accent

StylesPlaceholder: added all themes to divs, so they get exported right

Finished ThemeSelector component: now it correctly changes the theme

Track: now the text of the tracks is in text-accent color

TrackList: corrected text colors;
Added state management to the jammmName variable.
Improved the map of the tracks: the first track does not show the up button, and the last one does not show the down button.
Buttons now have tooltips