Exported calculateTotalTime
SearchResultsContainer.jsx: removed async keyword before sortTracks function, because it is not needed

calculateTotalTime.js: created a "library" to hold some methods and, in this case, it holds the method to caluculta the total time of the selected tracks that go in the tracklist,

TrackList.jsx: since the calculateTotalTime method is now in another file, it does not need to be in this one. The method now gets imported and is used when eeded.
Also, added some border to the div that holds the tracks