Finished logics
page.js: changed alert message to a toast

SearchResultsContainer.jsx: finished tracklist manipulation logic. Now tracks can be moved up, down, or removed.
Created submit form logic, which is passed as a prop to TrackList.jsx
This method, for now, only "sends" the request to create the playlist.

Tracklist.jsx: this component now recieves the submitForm method from its parent and handles it with a local method, to send the jammmName to the parent.

Added the spotifyApis.js lib, which will hold the methods to request the searched string and then to create the playlist.