Update logic
Finished the logic!

page.js: passes down the setJammmUri method so that SearchResultsContainer can update it and, by doing so, a new panel is shown with a button that opens a new tab/window directly to the newly created playlist.

Button.jsx: added a new prop: customClass to better stylize it

JammmLink.jsx: it recieves the jammmUri variable as prop and shows a button that once clicked opens a new tab/window to the playlist

Rules.jsx: renamed it to JammmLink.jsx

SearchResultsContainer.jsx: recieves a new prop to update the jammmUri when the submitForm is finished

ThemeSelectorContainer.jsx: it now selects Retro as the default theme