Update toasts
page.js: changed toast.warning to a toast.loading, which is updated after the sarch is done

SearchPagination.jsx: changed from flex to grid, and added a placeholder in each of the buttons. This keeps the buttons in the same place, even if the other button does not appear

SearchResultsContainer.jsx: changed the toasts to toast.loading, which are updated after the changes are complete

TrackList.jsx: wrapped the buttons in a div with grid layout. Added some transparent placeholders to when one of the buttons is not shown. This keeps the layout consistent and the buttons aligned at all times

calculateTotalTime.js: removed comments

spotifyApis.js: removed comments